Move paddle dimensions from view to Game
Take dimensions of ball into account when moving the ball.

Before this commit, the ball was simulated as an extension-less point,
and it was sufficient to add a margin to the board to not have the ball
clipped.

This commit is a preparation for the addition of a WebGL backend. When a
Game contains the complete geometry that is needed for rendering the
board, the WebGL backend becomes easier to implement.